Combine specific Random instance tests in one file.
